refactor: - RacingCarGame: 내부에서 ResultView 사용하는 부분 제거 - RacingGamePlayer: stage 별 결과 출력하는 부분 역할 추가
