+++ src/main/java/racingcar/racing/view/ResultView.java
@@ -7,13 +7,13 @@ import java.util.List;
 public class ResultView {
     private static final String RESULT_POSTFIX = "가 최종 우승했습니다.";
 
-    public void printGameStatus(List<Car> cars, int stage) {
+    public static void printGameStatus(List<Car> cars, int stage) {
         System.out.println("stage: " + stage);
         cars.forEach(car -> System.out.printf("%-5s : %s%n", car.getCarName(), "-".repeat(car.getPosition())));
         System.out.println();
     }
 
-    public void printResult(List<String> winners) {
+    public static void printResult(List<String> winners) {
         System.out.printf("%s%s\n", String.join(", ", winners), RESULT_POSTFIX);
     }
 }
